Move repository factor outside 'db'
