+++ api/src/index.spec.ts
@@ -6,6 +6,7 @@ import sinon from 'sinon';
 import { Connection, getConnection } from 'typeorm';
 import { createDatabaseConnection } from '@/server/db';
 import { MainScheduler, redis as queueRedis } from '@/server/queue';
+import { RepositoryFactory } from '@/server/repository-factory';
 import { redis } from '@/server/queue/redis';
 import { envVars } from '@/server/core/env-vars';
 
@@ -14,9 +15,10 @@ chai.should();
 
 before(async () => {
   await createTestDatabase();
-  await createDatabaseConnection({
+  const connection = await createDatabaseConnection({
     migrationsRun: true,
   });
+  RepositoryFactory.setDefaultConnection(connection);
 });
 
 beforeEach(async () => {
