+++ api/src/server/db/index.ts
@@ -2,9 +2,6 @@ import { Connection, createConnection, ConnectionOptions } from 'typeorm';
 import ormconfig from '@/../ormconfig';
 
 import { UniversePropertyRepository } from './repository/game/universe-property-repository';
-import { RepositoryFactory } from './repository-factory';
-
-export { RepositoryFactory };
 
 type ReposType = {
   UniversePropertyRepo: UniversePropertyRepository;
@@ -21,8 +18,6 @@ export async function createDatabaseConnection(
     ...ormconfig,
     ...options,
   } as ConnectionOptions);
-
-  RepositoryFactory.setDefaultConnection(connection);
 
   UniversePropertyRepo = connection.getCustomRepository(
     UniversePropertyRepository
