+++ api/src/server/repository-factory.ts
@@ -1,9 +1,8 @@
 import { Connection, EntityManager } from 'typeorm';
 import { identifierProvider } from '@/server/plugins/shared/uuid-identifier-provider/identifier-provider';
 import { raceKeyProvider } from '@/server/plugins/game/races/available-race-provider';
-
-import { createEntityCharacterRepository } from './repository/game/character.repository';
-import { createAccountRepository } from './repository/access/account.repository';
+import { createEntityCharacterRepository } from '@/server/db/repository/game/character.repository';
+import { createAccountRepository } from '@/server/db/repository/access/account.repository';
 
 export class RepositoryFactory {
   private static DefaultConnectionInstance: RepositoryFactory;
